_ #2 adds meta program run.py to manage projects
run.py allows users creating a topicflow project in terminal and see the visualization in a local server.
Five main functions are defined for modifying TopicFlow scripts and data transformation.
A jupyter notebook is added to explain how TopicFlow for PERCEIVE and run.py works in detail.
Several changes are made to index.html and controller.js to modify the user interface showing only the PERCEIVE related content.
All irrelevant data files are also removed.
A backup file is put in the main directory in case something went wrong so that the user can always revert the tool to the original state.