_ #2 adds os portability and run.py interface
Now utilizes os.path.join to support both Windows and unix styled path objects.
argparser is updated from (--run, --new, --path) to only (--add), three arguments are required to add a new project.
a checker is added to validate the two paths.
The notebook and README are updated to reflect the changes.