_ #2 folder organization, delete function, new run.py interface
Functions modify_html and modify_controller are moved after the data transformation functions so that index.html and controller.js stays are unchanged even though the data transformation breaks.
A new delete function is added to delete existing project(s).
Five arguments are required to add a new project (added doc_extension and path of document metadata).
The architecture of transform_doc and transform_bins are changed for more robustness.
The notebook and README are updated to reflect the changes.
A new folder is added to store both .xml and .png files of all graphs used in the notebook.
The backup folder is updated containing only the necessary files.